pyproject.toml
[project]
name = "seo-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "google-api-python-client>=2.193.0",
    "google-auth>=2.49.1",
    "google-auth-oauthlib>=1.3.1",
    "httpx>=0.28.1",
    "langchain-core>=1.2.26",
    "langchain-google-genai>=4.2.1",
    "langgraph>=1.1.6",
    "lxml>=6.0.2",
    "python-dotenv>=1.2.2",
    "rich>=14.3.3",
]
